Move files back to arrow package.
